added visitRequest pojo
