+++ SquidSquAds/src/main/java/com/squidsquads/controller/VisitController.java
@@ -1,15 +1,13 @@
 package com.squidsquads.controller;
 
+import com.squidsquads.form.visit.request.VisitRequest;
 import com.squidsquads.form.visit.response.CookieCreationResponse;
 import com.squidsquads.form.visit.response.VisitResponse;
 import com.squidsquads.service.visit.VisitService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
@@ -32,8 +30,8 @@ public class VisitController {
     }
 
     @PostMapping("")
-    public ResponseEntity<CookieCreationResponse> createIdentity(HttpServletResponse httpResponse) {
-        CookieCreationResponse response = visitService.createIdentity();
+    public ResponseEntity<CookieCreationResponse> createIdentity(HttpServletResponse httpResponse, @RequestBody VisitRequest request) {
+        CookieCreationResponse response = visitService.createIdentity(request);
         Cookie cookie = new Cookie("_squidsquads", response.getFingerprint());
         cookie.setMaxAge(2678400); // 1 mois
         httpResponse.addCookie(cookie);
